pytestui-mob-example-main/requirements.txt
python-testui
behave
pylint
allure-behave